Schematics 01-03
Schematic #03 might need to be slightly altered, but for the most part
these jpegs are okay. I won't add them into the report till all are
done. The numbering convention follows the list that is in the lab4
assignment except they didn't number the list, I am numbering from top
to bottom starting at zero, so we should be able to see clearly what is
done and not done.