Bruce ver1
This is what I have so far. I decided it would be nice to have a flow
chart of the program for our benefit and so we could also possibly put
it in the report. I don't know where it would go but most likely it
would be in the software section. I am still working on the report but
before I continue I would like to meet to discuss some of these
sections.